requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-options --output-file=requirements.txt requirements.in
#
async-timeout==4.0.2
    # via redis
certifi==2022.6.15
    # via sentry-sdk
click==8.1.3
    # via rq
deprecated==1.2.13
    # via redis
packaging==21.3
    # via redis
pyparsing==3.0.9
    # via packaging
python-dotenv==0.20.0
    # via -r requirements.in
redis==4.3.4
    # via
    #   -r requirements.in
    #   rq
rq==1.10.1
    # via -r requirements.in
sentry-sdk==1.7.2
    # via -r requirements.in
urllib3==1.26.10
    # via sentry-sdk
wrapt==1.14.1
    # via deprecated
